requirements.txt
minio==7.1.16
pandas==2.1.0
scikit-learn==1.3.0
joblib==1.3.2
pyyaml==6.0.1
apache-airflow-providers-amazon==8.7.0